requirements.txt
streamlit
streamlit-autorefresh
extra-streamlit-components
bcrypt
itsdangerous
pandas
python-dotenv
APScheduler
